% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{do_one_timestep}
\alias{do_one_timestep}
\title{do_one_timestep}
\usage{
do_one_timestep(Children_matr, Adults_matr, Cmatrix, Beta_par, Gamma_par)
}
\description{
do_one_timestep
}
\keyword{external}
