% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{do_one_timestep_bis}
\alias{do_one_timestep_bis}
\title{do_one_timestep_bis}
\usage{
do_one_timestep_bis(Children_matr, Adults_matr, Cmatrix, Beta_par,
  Gamma_par)
}
\description{
do_one_timestep_bis
}
\keyword{external}
