% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{do_one_timestep_tris}
\alias{do_one_timestep_tris}
\title{do_one_timestep_tris}
\usage{
do_one_timestep_tris(Children_matr, Adults_matr, Cmatrix, Beta_par,
  Gamma_par, list_trav_ch, list_trav_ad)
}
\description{
do_one_timestep_tris
}
\keyword{external}
