% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{generate_travel_uniform}
\alias{generate_travel_uniform}
\title{generate_travel_uniform}
\usage{
generate_travel_uniform(matr_pop, trav_prob)
}
\description{
generate_travel_uniform
}
\keyword{external}
