% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_geo_plot_functions.R
\name{geo_plot_social_contact_radius}
\alias{geo_plot_social_contact_radius}
\title{Plot the population and focus on the social contact radius}
\usage{
geo_plot_social_contact_radius(pop_data, area_size, max_contact_distance,
  average_num_contacts)
}
\arguments{
\item{pop_data}{the vector with population data}

\item{area_size}{the total area size}

\item{max_contact_distance}{the max distance between 2 individuals for a contact}

\item{average_num_contacts}{the average number of contacts per day}
}
\description{
This function shows the spatial configuration of the population
with color codes for the health state and the social contact radious of one individual.
}
\keyword{external}
