% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_function_lib.R
\name{get_contact_probability}
\alias{get_contact_probability}
\title{Calculate the social contact probability}
\usage{
get_contact_probability(average_num_contacts, num_possible_contacts)
}
\arguments{
\item{average_num_contacts}{the average number of contacts per time step}

\item{num_possible_contacts}{the number of possible contacts at this time step}
}
\description{
This function calculates the social contact probability based on
the average number of contacts per time step and the number of possible
social contacts at this time step.
}
\note{
The maximum probability is limited to 0.999
}
\keyword{external}
