% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ibm_sirv_geo.R
\name{get_default_param_ibm_sirv_geo}
\alias{get_default_param_ibm_sirv_geo}
\title{Get de default model parameters for one ibm_sirv_geo tutorial simulation}
\usage{
get_default_param_ibm_sirv_geo()
}
\description{
Returns de default model parameters of the ibm_sirv_geo tutorial.
}
\keyword{external}
