% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population_matrix.R
\name{get_default_population_matrix}
\alias{get_default_population_matrix}
\title{Load a precomputed synthetic population with households}
\usage{
get_default_population_matrix(pop_size, area_size)
}
\arguments{
\item{pop_size}{the final population size}

\item{area_size}{the size of the simulation area (to sample coordinates)}
}
\description{
Load a precomputed synthetic population with households
}
\keyword{external}
