% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot_functions.R
\name{plot_global_epidemic_ageclasses}
\alias{plot_global_epidemic_ageclasses}
\title{plot_global_epidemic_ageclasses}
\usage{
plot_global_epidemic_ageclasses(list_results)
}
\description{
plot_global_epidemic_ageclasses
}
\keyword{external}
