% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot_functions.R
\name{plot_infected_ageclasses}
\alias{plot_infected_ageclasses}
\title{plot_infected_ageclasses}
\usage{
plot_infected_ageclasses(list_results, xM)
}
\description{
plot_infected_ageclasses
}
\keyword{external}
