% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_plot_functions.R
\name{plot_infected_ageclasses_multipleSims}
\alias{plot_infected_ageclasses_multipleSims}
\title{plot_infected_ageclasses}
\usage{
plot_infected_ageclasses_multipleSims(list_all_results, Nsimulations)
}
\description{
plot_infected_ageclasses
}
\keyword{external}
