% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_function_lib.R
\name{plot_pop_health_states}
\alias{plot_pop_health_states}
\title{Plot of the population by health state}
\usage{
plot_pop_health_states(pop_data, area_size, i_day, plot_time_delay)
}
\arguments{
\item{pop_data}{the vector with population data}

\item{area_size}{the total area size}

\item{i_day}{the current day}

\item{plot_time_delay}{the time delay between two plots}
}
\description{
This function shows the spatial configuration of the population
with color codes for the health state and tracks one individual.
}
\keyword{external}
