% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_random_walk.R
\name{plot_social_contact_radius}
\alias{plot_social_contact_radius}
\title{Visualise the social contact radius}
\usage{
plot_social_contact_radius(pop_data, area_size, contact_distance,
  num_contacts_day)
}
\arguments{
\item{pop_data}{the vector with population data}

\item{area_size}{the total area size}

\item{contact_distance}{the current day}

\item{num_contacts_day}{the time delay between two plots}
}
\description{
This function shows the spatial configuration of the population
with color codes for the health state and the social contact radious of one individual.
}
\keyword{external}
