% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_function_lib.R
\name{print_sirv_geo_param}
\alias{print_sirv_geo_param}
\title{Print model parameters from the ibm_sirv_geo tutorial on the console}
\usage{
print_sirv_geo_param()
}
\description{
This functions provides an overview of the model parameters
}
\note{
This function is created for the ibm_sirv_geo tutorial
}
\keyword{external}
