% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ibm_sirv_geo.R
\name{run_ibm_sirv_geo}
\alias{run_ibm_sirv_geo}
\title{Run the ibm_sirv_geo tutorial with given model settings}
\usage{
run_ibm_sirv_geo(sim_param = NULL)
}
\arguments{
\item{sim_param}{Model parameters for this simulation}
}
\description{
Returns de default model parameters of the ibm_sirv_geo tutorial.
}
\keyword{external}
