% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{run_single_sim}
\alias{run_single_sim}
\title{run_single_sim}
\usage{
run_single_sim(NPatches, NTimes, ContactMatrix, Beta, Gamma,
  PopulationChildren, PopulationAdults, InitialInfected_ch,
  InitialInfected_ad)
}
\description{
run_single_sim
}
\keyword{external}
