% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm_function_lib.R
\name{sample_vaccine_refusal}
\alias{sample_vaccine_refusal}
\title{EXAMPLE to incorporate spatial vaccine refusal}
\usage{
sample_vaccine_refusal(pop_data, vaccine_coverage)
}
\arguments{
\item{vaccine_coverage}{the vaccine coverage}

\item{pop_size}{matrix with population data}
}
\description{
This functions assumes spatial vaccine refusal in the
outer regions of the simulated area.
}
\keyword{external}
