% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_package_admin.R
\docType{package}
\name{simid.course.2019}
\alias{simid.course.2019}
\alias{simid.course.2019-package}
\title{simid.course.2019:  Material for SIMID course 2019}
\description{
simid.course.2019:  Material for SIMID course 2019
}
\author{
Lander Willem [lander.willem@uantwerpen.be]
}
\keyword{internal}
