% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{spread_in_patch_and_among_travellers}
\alias{spread_in_patch_and_among_travellers}
\title{spread_in_patch_and_among_travellers}
\usage{
spread_in_patch_and_among_travellers(matrPopC, matrPopA, beta_par,
  gamma_par, ContMatr, Ch_S_trav, Ch_I_trav, Ch_R_trav, Ad_S_trav,
  Ad_I_trav, Ad_R_trav)
}
\description{
spread_in_patch_and_among_travellers
}
\keyword{external}
