% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{traveler_data_ER}
\alias{traveler_data_ER}
\title{traveler_data_ER}
\usage{
traveler_data_ER(prob_link, Npatches, prob_travel, reference_population)
}
\description{
traveler_data_ER
}
\keyword{external}
