% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{traveler_data_fully_connected}
\alias{traveler_data_fully_connected}
\title{traveler_data_fully_connected}
\usage{
traveler_data_fully_connected(Npatches, prob_travel, reference_population)
}
\description{
traveler_data_fully_connected
}
\keyword{external}
