% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population_matrix.R
\name{create_population_matrix}
\alias{create_population_matrix}
\title{Create a synthetic population with households of the given size}
\usage{
create_population_matrix(pop_size, area_size)
}
\arguments{
\item{pop_size}{the final population size}

\item{area_size}{the size of the simulation area (to sample coordinates)}
}
\description{
This function creates a population with households of size 4.
}
\keyword{external}
